+++ pkg/credentialcache/cache.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"syscall"
 )
 
 const (
@@ -62,31 +61,6 @@ func Read(issuerURL, clientID string) (Entry, error) {
 func lockFilePath(issuerURL, clientID string) string {
 	hash := sha256.Sum256([]byte(fmt.Sprintf("%s:%s", issuerURL, clientID)))
 	return filepath.Join(cacheDir(), fmt.Sprintf("token-%x.lock", hash[:16]))
-}
-
-func lock(issuerURL, clientID string) (*os.File, error) {
-	lockPath := lockFilePath(issuerURL, clientID)
-
-	if err := os.MkdirAll(filepath.Dir(lockPath), 0700); err != nil {
-		return nil, err
-	}
-
-	f, err := os.OpenFile(lockPath, os.O_CREATE|os.O_RDWR, 0600)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := syscall.Flock(int(f.Fd()), syscall.LOCK_EX); err != nil {
-		_ = f.Close()
-		return nil, err
-	}
-
-	return f, nil
-}
-
-func unlock(f *os.File) {
-	_ = syscall.Flock(int(f.Fd()), syscall.LOCK_UN)
-	_ = f.Close()
 }
 
 func filePath(issuerURL, clientID string) string {